wspgstat.sh: Enhance timestamp accuracy
In wspgstat.sh, timestamp is calculated before getting number of pages
in working set.  Because the process for the number of pages requires
multiple seconds for large size of working set, the timestamp in final
output has low accuracy.  This commit fixes the problem by calculating
timestamp after getting number of pages in working set.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>